+++ Snabble/Core/API/APIVersion.swift
@@ -5,5 +5,5 @@
 //
 
 enum APIVersion {
-    static let version = "0.10.3"
+    static let version = "0.10.4"
 }