+++ Snabble/UI/Payment/PaymentMethodSelectionViewController.swift
@@ -124,7 +124,7 @@ final class PaymentMethodSelectionViewController: UIViewController {
             case .success(let viewController):
                 self.navigationController?.pushViewController(viewController, animated: true)
             case .failure(let error):
-                let handled = self.process.delegate.handlePaymentError(error)
+                let handled = self.process.delegate.handlePaymentError(method, error)
                 if !handled {
                     self.process.delegate.showWarningMessage("Snabble.Payment.errorStarting".localized())
                 }